refactor[#34]: 가중치 전략 패턴을 WeightManager로 대체
