--- src/main/java/org/highfive/backend/global/strategy/weight/GenreWeightStrategy.java
@@ -1,9 +1,0 @@
-package org.highfive.backend.global.strategy.weight;
-
-import java.util.List;
-import java.util.Map;
-
-public interface GenreWeightStrategy {
-
-    Map<String, Double> calcWeight(Map<String, Integer> genreCount);
-}
